[refactor] Expose getParam to be used on other plugins
This is required by the ep_script_scene_marks
https://github.com/storytouch/ep_script_scene_marks/pull/51
